RAPyDo 3.0 (#504)
## :star2: New functionalities

Alchemy connector is now based on pure SQLAlchemy (dropped support for Flask-SQLAlchemy)
Replace obsolete passlib with direct use of bcrypt
Add Alembic to handle sql migrations
Set broker_connection_retry_on_startup to celery conf
Add SMTP_REPLYTO env variable
Add retry option on connectors get_instance

## :bug: Bug fixes and improvements

Replaced legacy sqlalchemy backref with back_populates
Cleanup of connectors.load_models method
Extended sqlalchemy models with Mapped types
Improve models definitions by using DeclarativeBase and mapped_column
Set alchemy READ COMMITTED isolation_level
Added session commits on sql statements to reduce the number of idle in transaction queries
Add __repr__ method to models classes to fix caching issues
Add neobolt to dependencies
Added SMTPConnectError to SMTP connection errors
Fix alchemy connection management to have a global engine
Fix to re-enable alchemy poolsize
Fix json encoder
Fix restapi init
Fix db initialization
Fix neo4j models by specifying full paths
Fix requirements.types to reduce vulnerabilities
Fix password reset
Bug fix to invert user-group relationship
Bug fix to prevent groups to be deleted with users still assign to
Bug fix when sending cc and bcc emails as a list
Extended flake8 configuration with flake8-bugbear
Add tests on legacy backends
Add pytest-rerunfailures to restart flaky tests
Enable semgrep analysis
Merge static analysis tools into a single workflow
Add psycopg2 types

## :wastebasket: Dropped functionalities

Drop Flask-SQLAlchemy in favor of pure SQLAlchemy
Drop sqlalchemy-stubs
Drop support for Flask-Migrate
Bump minimum python version to 3.9 (drop support for py38)
Drop legacy38 backend build
Drop ggshield hook
Remove obsolete filterwarnings
Remove unsupported encoding from alchemy create_engine

## :rocket: Dependencies Upgrades

Bump alembic to 1.9.4
Bump amqp to 5.2.0
Bump apispec to 6.3.1
Bump celery to 5.4.0
Bump celery-redbeat to 2.2.0
Bump Flask to 2.2.5
Bump Flask-Caching to 2.1.0
Bump Flask-Cors to 4.0.0
Bump flower to 2.0.1
Bump glom to 23.5.0
Bump html2text to 2024.2.26
Bump loguru to 0.7.2
Bump lxml to 5.2.1
Bump marshmallow to 3.21.1
Bump neomodel to 5.2.1
Bump orjson to 3.10.1
Bump pika to 1.3.2
Bump plumbum to 1.8.2
Bump psutil to 5.9.8
Bump psycopg2-binary to 2.9.9
Bump PyJWT to 2.8.0
Bump PyMySQL to 1.1.0
Bump pyOpenSSL to 23.0.0
Bump pyotp to 2.9.0
Bump PyYAML to 6.0.1
Bump redis to 5.0.4
Bump requests to 2.31.0
Bump segno to 1.6.1
Bump sentry to 2.0.1
Bump SQLAlchemy to 2.0.29
Bump webargs to 8.4.0
Bump Werkzeug to 3.0.2
Bump pytest to 8.1.1
Bump pytest-cov to 5.0.0
Bump pytest-flask to 1.3.0
Bump pytest-sugar to 1.0.0
Bump pytest-timeout to 2.3.1
Bump Faker to 24.9.0
Bump celery-types to 0.22.0
Bump types-python-dateutil to 2.9.0.20240316
Bump types-pytz to 2024.1.0.20240417
Bump types-PyYAML to 6.0.12.20240311
Bump types-redis to 4.6.0.20240425
Bump types-requests to 2.31.0.20240406
Bump types-urllib3 to 1.26.25.14
Bump pre-commit hooks
Bump actions/setup-python action to v5
Bump upload-artifact action to v4
Bump checkout action to v4
Bump codeql-action action to v3
Bump black to 23.1.0

+++ restapi/endpoints/admin_groups.py
@@ -1,15 +1,13 @@
-from typing import Any, Dict, List
+from typing import Any
 
 from restapi import decorators
-from restapi.connectors import Connector
 from restapi.endpoints.schemas import GroupWithMembers, admin_group_input
-from restapi.exceptions import NotFound
+from restapi.exceptions import Forbidden, NotFound
 from restapi.rest.definition import EndpointResource, Response
 from restapi.services.authentication import Group, Role, User
 
 
-def inject_group(endpoint: EndpointResource, group_id: str) -> Dict[str, Any]:
-
+def inject_group(endpoint: EndpointResource, group_id: str) -> dict[str, Any]:
     group = endpoint.auth.get_group(group_id=group_id)
     if not group:
         raise NotFound("This group cannot be found")
@@ -18,7 +16,6 @@ def inject_group(endpoint: EndpointResource, group_id: str) -> Dict[str, Any]:
 
 
 class AdminGroups(EndpointResource):
-
     depends_on = ["AUTH_ENABLE"]
     labels = ["management"]
     private = True
@@ -34,12 +31,10 @@ class AdminGroups(EndpointResource):
         },
     )
     def get(self, user: User) -> Response:
-
-        groups: List[Dict[str, Any]] = []
+        groups: list[dict[str, Any]] = []
 
         is_admin = self.auth.is_admin(user)
         for g in self.auth.get_groups():
-
             members = list(g.members)
 
             coordinators = [
@@ -73,7 +68,6 @@ class AdminGroups(EndpointResource):
         },
     )
     def post(self, user: User, **kwargs: Any) -> Response:
-
         group = self.auth.create_group(kwargs)
 
         self.auth.save_group(group)
@@ -91,7 +85,6 @@ class AdminGroups(EndpointResource):
         responses={204: "Group successfully modified", 404: "Group not found"},
     )
     def put(self, group_id: str, group: Group, user: User, **kwargs: Any) -> Response:
-
         # mypy correctly raises errors because update_properties is not defined
         # in generic Connector instances, but in this case this is an instance
         # of an auth db and their implementation always contains this method
@@ -111,6 +104,10 @@ class AdminGroups(EndpointResource):
         responses={204: "Group successfully deleted", 404: "Group not found"},
     )
     def delete(self, group_id: str, group: Group, user: User) -> Response:
+        if members := self.auth.get_group_members(group):
+            raise Forbidden(
+                f"Cannot delete this group, it is assigned to {len(members)} user(s)"
+            )
 
         self.auth.delete_group(group)
